Conflicts:
	debian/changelog
	debian/control
	salt/__init__.py
	salt/master.py

+++ salt/log/handlers/logstash_mod.py
@@ -155,9 +155,7 @@
 
 # Import python libs
 import os
-import zmq
 import json
-import socket
 import logging
 import logging.handlers
 import datetime
@@ -166,6 +164,13 @@ import datetime
 from salt._compat import string_types
 from salt.log.setup import LOG_LEVELS
 from salt.log.mixins import NewStyleClassMixIn
+import salt.utils.network
+
+# Import Third party libs
+try:
+    import zmq
+except ImportError:
+    pass
 
 log = logging.getLogger(__name__)
 
@@ -267,7 +272,7 @@ class LogstashFormatter(logging.Formatter, NewStyleClassMixIn):
         return datetime.datetime.utcfromtimestamp(record.created).isoformat()[:-3] + 'Z'
 
     def format_v0(self, record):
-        host = socket.getfqdn()
+        host = salt.utils.network.get_fqhostname()
         message_dict = {
             '@timestamp': self.formatTime(record),
             '@fields': {
@@ -322,7 +327,7 @@ class LogstashFormatter(logging.Formatter, NewStyleClassMixIn):
         message_dict = {
             '@version': 1,
             '@timestamp': self.formatTime(record),
-            'host': socket.getfqdn(),
+            'host': salt.utils.network.get_fqhostname(),
             'levelname': record.levelname,
             'logger': record.name,
             'lineno': record.lineno,
@@ -413,4 +418,12 @@ class ZMQLogstashHander(logging.Handler, NewStyleClassMixIn):
     def close(self):
         if self._context is not None:
             # One second to send any queued messages
-            self._context.destroy(1 * 1000)
+            if hasattr(self._context, 'destroy'):
+                self._context.destroy(1 * 1000)
+            else:
+                if getattr(self, '_publisher', None) is not None:
+                    self._publisher.setsockopt(zmq.LINGER, 1 * 1000)
+                    self._publisher.close()
+
+                if self._context.closed is False:
+                    self._context.term()
